requirements_simple.txt
# Минимальные зависимости для простого тестового бота
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
